// frontend/src/app/api/model/scanStatus.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * SendR API
 * WeTransfer-like file sharing service
 * OpenAPI spec version: 0.1.0
 */

export type ScanStatus = typeof ScanStatus[keyof typeof ScanStatus];


export const ScanStatus = {
  queued: 'queued',
  scanning: 'scanning',
  clean: 'clean',
  infected: 'infected',
  failed: 'failed',
} as const;